WIP: get simple example from mpstthree working
mpstthree is a library for multiparty session types made by the Imperial
mobility group.
Three here as far as I can see means that we can have three
communicating parties.
It seems there is some work going on to try to extend this to more than
three.

We are trying to model the interraction Kernel<->ServerApp<->ClientApp
so three communicating participants should be fine.